Move GenAI plugins to dedicated directory
