+++ frigate/genai/plugins/gemini.py
