+++ frigate/genai/plugins/llama_cpp.py
