+++ frigate/genai/plugins/ollama.py
