+++ frigate/genai/plugins/openai.py
